requirements.txt
solara
python-pptx
pytest
